Export functions necessary for ebpf-go port (#4116)
* Export ebpf_close_fd and ebpf_dup_fd

Export functions used to manipulate fd_t. This makes the use of the UCRT
_get_osfhandle an implementation detail which is abstracted away from
users of ebpfapi.dll. This also makes it easier to use the API via
run-time dynamic linking.

* Allow attaching a link as an fd

Add a function which returns the link as an fd when attaching a program.
The existing functions are re-jigged to reduce code duplication.

* Export functions to translate between GUID and enums

Allow users to resolve libbpf style enums to Windows GUIDs and vice
versa.

* Fix up spelling mistakes

Co-authored-by: Dave Thaler <dthaler1968@gmail.com>

* Fix sorting of Source.def

* Rename to ebpf_program_attach_fds

* Add tests for ebpf_dup_fd and ebpf_close_fd

* Rename link argument to link_handle

* Document return value of ebpf_program_attach_fds

Co-authored-by: Dave Thaler <dthaler1968@gmail.com>

* Fix ebpf_get_attach_type use-after-free

* Rename ebpf_dup_fd to ebpf_duplicate_fd

* Rename to ebpf_program_attach_by_fds

* Fix docstring of ebpf_get_ebpf_attach_type

* Temporarily disable _wer_report::_terminate_hook

---------

Co-authored-by: Dave Thaler <dthaler1968@gmail.com>